Example Fast Track aggregated data

An example of aggregated data as provided by Fast Track (in an 'aggregated_data.csv' file). The data is productions of 12 vowels by 7 adult males from the Hillenbrand et al. (1995) data. The formant columns are so that column FXY represents the frequency for the Xth formant, for the Yth time slice.

aggregated_data

Format

An object of class data.frame with 72 rows and 22 columns.

Source

https://github.com/santiagobarreda/FastTrackR

Columns

file

the recording file name.

f0

average f0 in Hertz.

duration

vowel duration in milliseconds.

label

The label used for plotting the vowel.

group

A group number, used for plotting

color

The color Praat will use for plotting.

number

The file number.

f11

The average frequency for the first formant, for the first time slice.

f21

The average frequency for the second formant, for the first time slice.

...

And so on.

References

Hillenbrand, J.M., Getty, L.A., Clark, M.J., and Wheeler, K. (1995). "Acoustic characteristics of American English vowels," Journal of the Acoustical Society of America, 97, 3099-3111.