Aggregate analysis data

Aggregate the .csv files contained in the csvs folder in a Fast Track directory. These files contain formant measurements at every 2ms, and there is one file per vowel token, so this function summarizes that data into a single dataframe with one row per token and measurements at a specified number of bins.

aggregatedata(path, csvs, bins = 5, n_formants = NA, method = "median")

Arguments

path
a string. The path to the working directory for the Fast Track project. If no path is provided, the current working directory for the current R session is used.#'

csvs
An object from the output of readcsvs().

bins
an integer. How many timepoints do you want formant data from? By default, you'll get formant samples at five points along the duration of each vowel.

n_formants
an integer. By default, aggregatedata will use the number of formants as is contained in csvs or in the .csv files. However, if you want to, for example, only aggregated data from F1, F2, and F3 even though you have data from F4, you can do so by setting n_formants to 3.

method
a string. Median by default. Other functions to come later.

Value

A dataframe containing formant measurements and various other information for each file (=vowel token)

Details

Note that if a file called aggregated_data.csv exists in the processed_data directory, this function will read that in instead—but only if the csvs object is not specified. If the file exists, but you want to process the data yourself (say, into 11 bins instead of 5), then include the csvs object in the function call. See the examples below.

Examples

if (FALSE) {
path <- "path/to/fasttrack/data"

# Read in the aggregated data
aggregatedata(path)

# Aggregate the csv files. This generates a spreadsheet identical to the one produced by Praat.
csvs <- readcsvs(path)
aggregatedata(path, csvs)

# Reprocess existing csv data. Let's say that when I first analyzed the audio in Praat using Fast Track, I only had the data binned into 5 timepoints. Now, I want 11 timepoints, so I'll generate a new version of the aggregated data here.
aggregatedata(path, csvs, bins = 11)

# Only process the first three formants even though four are in the original csvs.
aggregatedata(path, csvs, n_formants = 3)
}