Normalize aggregated data

Normalize formant frequencies using the log mean method outlined in Barreda and Nearey (2018). This approach is robust to unbalanced and missing data and is appropriate to use with formant contours. If you do not provide a vector of talkers, the data is assumed to be produced by a single talker.

normalize(aggregated_data, talker = NA)

Arguments

aggregated_data
A dataframe representing data aggregated by Fast Track.

talker
an optional vector indicating which talker produced each row in the data.

Value

Another dataframe with normalized formant frequencies.

References

Barreda, S., & Nearey, T. M. (2018). A regression approach to vowel normalization for missing and unbalanced data. The Journal of the Acoustical Society of America, 144(1), 500-520.

Examples

data (aggregated_data)
normalize (aggregated_data)

#>     file    f0 duration label group   color number        f11          f21
#> 1  m01ae 169.3    0.322    ae     1    Blue      1 -0.6015350  0.477717034
#> 2  m01ah 161.7    0.316    ah     2   Green      2 -0.4162679  0.076004275
#> 3  m01aw 177.2    0.330    aw     3 Magenta      3 -0.6104240 -0.194105262
#> 4  m01eh 173.5    0.264    eh     4   Black      4 -0.7006874  0.464090782
#> 5  m01ei 167.9    0.318    ei     5    Lime      5 -0.8961239  0.536127796
#> 6  m01er 175.8    0.322    er     6  Purple      6 -0.8496939  0.119967399
#> 7  m01ih 178.0    0.266    ih     7    Teal      7 -1.1805110  0.562721386
#> 8  m01iy 175.4    0.338    iy     8    Navy      8 -1.2684798  0.662124967
#> 9  m01oa 170.0    0.368    oe     9    Pink      9 -0.8440388 -0.216935038
#> 10 m01oo 170.5    0.254    oo    10  Maroon     10 -1.0407491 -0.184339560
#> 11 m01uh 172.6    0.290    uh    11   Olive     11 -0.6284425 -0.025617918
#> 12 m01uw 186.6    0.280    uw    12    Grey     12 -1.1964265 -0.289808061
#> 13 m02ae 102.4    0.250    ae     1    Blue     13 -0.6421727  0.427346674
#> 14 m02ah 102.8    0.248    ah     2   Green     14 -0.4224142 -0.007540187
#> 15 m02aw 105.7    0.288    aw     3 Magenta     15 -0.6765510 -0.317177038
#> 16 m02eh 104.9    0.166    eh     4   Black     16 -0.8272630  0.419939233
#> 17 m02ei 106.0    0.260    ei     5    Lime     17 -0.9851174  0.507348831
#> 18 m02er 110.1    0.296    er     6  Purple     18 -0.9700795 -0.081898754
#> 19 m02ih 111.6    0.182    ih     7    Teal     19 -1.1143291  0.485207706
#> 20 m02iy 108.0    0.234    iy     8    Navy     20 -1.3065518  0.610809734
#> 21 m02oa 104.3    0.260    oe     9    Pink     21 -0.8515860 -0.295222259
#> 22 m02oo 109.3    0.174    oo    10  Maroon     22 -1.1045251 -0.257924382
#> 23 m02uh 101.4    0.172    uh    11   Olive     23 -0.6148982 -0.172745227
#> 24 m02uw 112.1    0.256    uw    12    Grey     24 -1.2858716 -0.253749011
#> 25 m03ae  92.8    0.344    ae     1    Blue     25 -0.7088640  0.363124392
#> 26 m03ah  97.4    0.372    ah     2   Green     26 -0.5221733 -0.109468308
#> 27 m03aw  90.6    0.354    aw     3 Magenta     27 -0.6089370 -0.283349481
#> 28 m03eh  94.5    0.282    eh     4   Black     28 -0.8365481  0.375970143
#> 29 m03ei  89.8    0.352    ei     5    Lime     29 -1.0025851  0.520882146
#> 30 m03er  93.1    0.368    er     6  Purple     30 -0.9468962 -0.033108590
#> 31 m03ih 101.6    0.260    ih     7    Teal     31 -1.0522567  0.452334961
#> 32 m03iy  93.6    0.314    iy     8    Navy     32 -1.2180490  0.626050612
#> 33 m03oa  95.7    0.416    oe     9    Pink     33 -0.8384155 -0.306139398
#> 34 m03oo  94.8    0.274    oo    10  Maroon     34 -0.9344737 -0.154658108
#> 35 m03uh  97.1    0.296    uh    11   Olive     35 -0.6639126 -0.156546683
#> 36 m03uw 101.4    0.418    uw    12    Grey     36 -1.1167952 -0.205951403
#> 37 m04ae 126.4    0.312    ae     1    Blue     37 -0.6749625  0.438877201
#> 38 m04ah 128.1    0.302    ah     2   Green     38 -0.4224142 -0.056778334
#> 39 m04aw 131.9    0.346    aw     3 Magenta     39 -0.5582382 -0.236195643
#> 40 m04eh 130.8    0.228    eh     4   Black     40 -0.6299588  0.360309867
#> 41 m04ei 135.9    0.252    ei     5    Lime     41 -0.8649323  0.564101648
#> 42 m04er 133.0    0.330    er     6  Purple     42 -0.7697966  0.141946305
#> 43 m04ih 133.3    0.226    ih     7    Teal     43 -0.9262769  0.464090782
#> 44 m04iy  87.8    0.238    iy     8    Navy     44 -1.1857881  0.653753215
#> 45 m04oa 146.7    0.300    oe     9    Pink     45 -0.8882343 -0.346458409
#> 46 m04oo 135.8    0.210    oo    10  Maroon     46 -0.8515860 -0.126749320
#> 47 m04uh 132.6    0.210    uh    11   Olive     47 -0.6030110 -0.113987069
#> 48 m04uw 143.3    0.288    uw    12    Grey     48 -1.0592254  0.037831702
#> 49 m06ae 115.6    0.252    ae     1    Blue     49 -0.6314774  0.401176957
#> 50 m06ah 114.3    0.228    ah     2   Green     50 -0.3730958  0.020562827
#> 51 m06aw 118.1    0.248    aw     3 Magenta     51 -0.5782103 -0.216935038
#> 52 m06eh 118.8    0.176    eh     4   Black     52 -0.7945328  0.388652971
#> 53 m06ei 117.9    0.240    ei     5    Lime     53 -0.8882343  0.472691898
#> 54 m06er 121.0    0.216    er     6  Purple     54 -0.8726394  0.098227412
#> 55 m06ih 113.4    0.158    ih     7    Teal     55 -0.9981895  0.470169828
#> 56 m06iy 122.8    0.240    iy     8    Navy     56 -1.1417965  0.601109481
#> 57 m06oa 116.8    0.262    oe     9    Pink     57 -0.7576542 -0.140606355
#> 58 m06oo 118.3    0.146    oo    10  Maroon     58 -0.9637033 -0.091594731
#> 59 m06uh 117.0    0.168    uh    11   Olive     59 -0.6670573 -0.062784358
#> 60 m06uw 123.1    0.230    uw    12    Grey     60 -1.1118691 -0.134115836
#> 61 m07ae  94.4    0.252    ae     1    Blue     61 -0.6877422  0.463582523
#> 62 m07ah  97.0    0.264    ah     2   Green     62 -0.4847357  0.084210215
#> 63 m07aw  96.2    0.290    aw     3 Magenta     63 -0.6360471 -0.240298213
#> 64 m07eh 105.4    0.196    eh     4   Black     64 -0.7456575  0.450276288
#> 65 m07ei 101.4    0.280    ei     5    Lime     65 -0.8017142  0.553005834
#> 66 m07er  96.4    0.272    er     6  Purple     66 -0.8272630  0.018978040
#> 67 m07ih 106.5    0.190    ih     7    Teal     67 -1.0875961  0.579160112
#> 68 m07iy  97.7    0.258    iy     8    Navy     68 -1.2207849  0.648696314
#> 69 m07oa  94.7    0.266    oe     9    Pink     69 -0.7439553 -0.014896166
#> 70 m07oo  91.8    0.196    oo    10  Maroon     70 -0.9938131 -0.065369452
#> 71 m07uh  93.3    0.184    uh    11   Olive     71 -0.6254167 -0.038133726
#> 72 m07uw 112.7    0.230    uw    12    Grey     72 -1.2262795 -0.144334225
#>          f31        f12          f22       f32        f13          f23
#> 1  0.7695259 -0.6178922  0.483215673 0.7699002 -0.6074522  0.471179419
#> 2  0.7661507 -0.4125982  0.081234022 0.7758704 -0.3849023  0.070747035
#> 3  0.6997572 -0.5610671 -0.180459826 0.7192371 -0.5400432 -0.154658108
#> 4  0.7710224 -0.6893512  0.454902356 0.7706485 -0.6623440  0.417280364
#> 5  0.7873371 -0.9262769  0.553005834 0.7994009 -0.9679496  0.586829899
#> 6  0.4476970 -0.8843127  0.116376731 0.3995523 -0.9040762  0.120683988
#> 7  0.8023037 -1.1778829  0.551144502 0.7902751 -1.0828114  0.522321340
#> 8  0.8596835 -1.2684798  0.670427215 0.8375449 -1.2742435  0.668772271
#> 9  0.7156875 -0.8745755 -0.224999598 0.7235584 -0.9020822 -0.245450208
#> 10 0.7021630 -0.9531658 -0.193124389 0.7168721 -0.9000921 -0.147139276
#> 11 0.7262987 -0.6208953 -0.004288151 0.7223817 -0.6239073  0.039386912
#> 12 0.6708405 -1.2071793 -0.299574823 0.6733168 -1.2180490 -0.307237696
#> 13 0.6745526 -0.6514320  0.412476512 0.7097433 -0.6421727  0.408724162
#> 14 0.8008533 -0.4860489 -0.044873432 0.7932044 -0.4768926 -0.015716847
#> 15 0.7356373 -0.6893512 -0.316067775 0.7506289 -0.6925770 -0.295222259
#> 16 0.6794808 -0.8309667  0.407649471 0.6737289 -0.7981171  0.375415050
#> 17 0.6997572 -1.0522567  0.518478876 0.7479546 -1.0972347  0.546943796
#> 18 0.2038077 -1.0407491 -0.047412578 0.1845059 -1.0685931 -0.006726186
#> 19 0.7329225 -1.0924038  0.470169828 0.6912911 -1.0638983  0.432079789
#> 20 0.9103779 -1.3155473  0.605971369 0.8986015 -1.3368571  0.612125235
#> 21 0.6969431 -0.9000921 -0.305042305 0.7125217 -1.0226080 -0.337357095
#> 22 0.5886260 -1.0638983 -0.215931526 0.5984475 -1.0407491 -0.150891626
#> 23 0.7684021 -0.6391052 -0.149013691 0.7650231 -0.6421727 -0.102280496
#> 24 0.5332869 -1.3035711 -0.273739156 0.5228006 -1.3035711 -0.277999013
#> 25 0.7196307 -0.7006874  0.365931018 0.7387309 -0.6813319  0.365931018
#> 26 0.7792130 -0.5046171 -0.069692848 0.7784712 -0.4847357 -0.029772473
#> 27 0.8156204 -0.6254167 -0.243386224 0.8149051 -0.6254167 -0.189210490
#> 28 0.6708405 -0.8071342  0.334037654 0.6592028 -0.7785609  0.316524071
#> 29 0.7414299 -1.0181237  0.534234752 0.7498655 -1.0615591  0.562721386
#> 30 0.2914780 -0.9658242 -0.010802832 0.2774918 -0.9872843 -0.003476792
#> 31 0.7313679 -1.0972347  0.432079789 0.7164774 -1.0780495  0.411405848
#> 32 0.8347428 -1.1937563  0.629501867 0.8364950 -1.1937563  0.629071111
#> 33 0.6925050 -0.9000921 -0.333965345 0.6937173 -0.9637033 -0.340760388
#> 34 0.7025634 -0.9510716 -0.156546683 0.6880470 -0.9448150 -0.134115836
#> 35 0.7180553 -0.6861358 -0.123086313 0.7344747 -0.6845319 -0.113081682
#> 36 0.6495409 -1.1342303 -0.206944945 0.6440382 -1.1545356 -0.211927516
#> 37 0.6989540 -0.6193926  0.416747741 0.6981501 -0.5400432  0.329978501
#> 38 0.7156875 -0.4089419 -0.011620160 0.7360245 -0.4236480 -0.005100169
#> 39 0.7880724 -0.5767704 -0.220959188 0.7888072 -0.5317558 -0.155601950
#> 40 0.7109350 -0.5610671  0.286028435 0.7125217 -0.5667489  0.291478039
#> 41 0.8949753 -0.9365334  0.609053036 0.9249060 -1.0615591  0.635084875
#> 42 1.0019857 -0.8272630  0.155182305 0.9894445 -0.8765154  0.170973914
#> 43 0.7464232 -0.8765154  0.404957680 0.7395028 -0.8440388  0.354089887
#> 44 1.0286305 -1.1910932  0.640636885 1.0239969 -1.2153204  0.638078249
#> 45 0.7081522 -0.9658242 -0.362587791 0.6896704 -1.0226080 -0.369580827
#> 46 0.7077541 -0.8162335 -0.069692848 0.6985522 -0.7963233  0.058625674
#> 47 0.7657750 -0.5386572 -0.038973709 0.7695259 -0.5331323  0.038609609
#> 48 0.6520704 -1.0948163  0.014208556 0.6541735 -1.1094151 -0.078396124
#> 49 0.7286415 -0.6406377  0.402258623 0.7180553 -0.6269285  0.386458784
#> 50 0.7305897 -0.3730958  0.010216535 0.7406595 -0.3707511  0.021354279
#> 51 0.7521539 -0.5442127 -0.203967275 0.7736358 -0.5483998 -0.206944945
#> 52 0.7464232 -0.7945328  0.377079405 0.7356373 -0.7874027  0.372078005
#> 53 0.7498655 -0.9573675  0.496585518 0.7498655 -0.9807977  0.517515949
#> 54 0.4011770 -0.9262769  0.081234022 0.3831585 -0.9510716  0.083466997
#> 55 0.7721432 -0.9959989  0.452848967 0.7623871 -0.9938131  0.439918867
#> 56 0.7784712 -1.1317209  0.625618367 0.8037519 -1.1443314  0.633370344
#> 57 0.6729045 -0.8125839 -0.196069899 0.6766090 -0.9262769 -0.233129724
#> 58 0.6864210 -0.9700795 -0.097814209 0.6937173 -0.9743530 -0.087175811
#> 59 0.7544370 -0.6845319 -0.051658869 0.7521539 -0.7187651 -0.054215325
#> 60 0.6359410 -1.1622576 -0.207939475 0.6389319 -1.2071793 -0.262117261
#> 61 0.7676522 -0.7287652  0.467641382 0.7290315 -0.6861358  0.435222798
#> 62 0.6687723 -0.4860489  0.076753059 0.6675293 -0.4926409  0.092349364
#> 63 0.6633747 -0.6329983 -0.224999598 0.6749642 -0.6074522 -0.165089687
#> 64 0.8252273 -0.7542118  0.428400415 0.7736358 -0.7803230  0.419408025
#> 65 0.8368451 -0.9262769  0.575530344 0.7758704 -1.0384634  0.607293247
#> 66 0.3070565 -0.9658242 -0.020655129 0.2651688 -1.0003849 -0.120347835
#> 67 0.8528177 -1.0522567  0.537545231 0.8098836 -0.9637033  0.504424853
#> 68 0.9357445 -1.1831461  0.648273749 0.9520970 -1.1805110  0.661708040
#> 69 0.6786611 -0.8804065 -0.150891626 0.6921005 -0.9829552 -0.213927517
#> 70 0.6908862 -0.9916321 -0.019830387 0.6819358 -0.9894558  0.029234541
#> 71 0.7262987 -0.6623440 -0.056778334 0.7471892 -0.6990600  0.034714006
#> 72 0.6419138 -1.2800406 -0.204958847 0.6465817 -1.2829518 -0.205951403
#>          f33        f14          f24       f34        f15         f25       f35
#> 1  0.7657750 -0.5724632  0.380952162 0.7601220 -0.7204249  0.37096318 0.7891744
#> 2  0.7570940 -0.4004621  0.103342513 0.7464232 -0.6733764  0.33750386 0.7702744
#> 3  0.7212036 -0.5126817 -0.016538203 0.7129179 -0.7645746  0.30883855 0.7247337
#> 4  0.7713961 -0.6781421  0.401717936 0.7713961 -0.8534817  0.40333912 0.8019413
#> 5  0.8023037 -1.1519747  0.620851294 0.8116799 -1.2401493  0.63636885 0.8209690
#> 6  0.3820559 -0.9468962  0.140542796 0.3754150 -1.1192674  0.25331956 0.4564396
#> 7  0.7773574 -0.9573675  0.451820690 0.7747537 -1.0615591  0.43574567 0.7975824
#> 8  0.8287621 -1.2656104  0.676608978 0.8213245 -1.2262795  0.67249205 0.8368451
#> 9  0.7231663 -0.9448150 -0.272677021 0.7180553 -1.2180490 -0.15560195 0.7101407
#> 10 0.7133140 -0.8862716  0.102613384 0.6945248 -0.9552644  0.32881571 0.7317568
#> 11 0.7223817 -0.6437099  0.134908979 0.7239503 -0.8591906  0.34325431 0.7699002
#> 12 0.6925050 -1.1964265 -0.266327793 0.6753757 -1.1622576 -0.05251030 0.6414883
#> 13 0.7029636 -0.6437099  0.372078005 0.6896704 -0.6941938  0.35579009 0.6860141
#> 14 0.7986739 -0.4652419  0.002990490 0.8051980 -0.4966171  0.08569500 0.7646469
#> 15 0.7601220 -0.6702119 -0.279066819 0.7555766 -0.6829306 -0.06278436 0.6989540
#> 16 0.6786611 -0.7838566  0.353522511 0.6749642 -0.8534817  0.34210686 0.6839770
#> 17 0.7695259 -1.1468727  0.579612908 0.7840216 -1.2713575  0.57598479 0.7747537
#> 18 0.1824878 -1.0804276  0.048667721 0.2031483 -1.1392681  0.13420251 0.2564514
#> 19 0.6716666 -1.0453363  0.396294929 0.6675293 -1.0476378  0.35465694 0.6691863
#> 20 0.8900092 -1.3185639  0.598891694 0.8830150 -1.3125398  0.58907457 0.8661626
#> 21 0.7113319 -1.0828114 -0.335094650 0.6495409 -1.1242302 -0.23619564 0.5953330
#> 22 0.5944413 -1.0271125 -0.061064667 0.6099318 -1.0407491  0.11781455 0.5737105
#> 23 0.7593659 -0.6592141 -0.014076158 0.7429690 -0.8089475  0.14684313 0.6957347
#> 24 0.5328126 -1.2858716 -0.221967761 0.5337609 -1.2771379 -0.09425549 0.5049128
#> 25 0.7559562 -0.6498828  0.352954813 0.7758704 -0.6797358  0.31475572 0.7494837
#> 26 0.7888072 -0.4939645 -0.015716847 0.7732628 -0.5456065  0.06014890 0.7033637
#> 27 0.8051980 -0.6269285 -0.159386250 0.7953958 -0.6514320 -0.02478907 0.6794808
#> 28 0.6704272 -0.7732931  0.287848271 0.6835691 -0.8107640  0.28237880 0.7196307
#> 29 0.7536765 -1.1443314  0.565939039 0.7448895 -1.1831461  0.57825390 0.7270802
#> 30 0.2744253 -1.0092150  0.007813650 0.2738108 -1.0361829  0.12211563 0.3141656
#> 31 0.7160825 -1.0545742  0.389200766 0.7156875 -1.0685931  0.35238679 0.7212036
#> 32 0.8241644 -1.2098857  0.613876549 0.7953958 -1.2044802  0.59844755 0.7468063
#> 33 0.6916959 -1.0384634 -0.337357095 0.7021630 -1.0924038 -0.25897096 0.6949282
#> 34 0.6945248 -0.9406656 -0.072295887 0.6786611 -0.9807977  0.15794665 0.6807090
#> 35 0.7471892 -0.7121535 -0.064507011 0.7329225 -0.7820882  0.12211563 0.7160825
#> 36 0.6495409 -1.1726473 -0.202976686 0.6342280 -1.1857881 -0.11308168 0.6103708
#> 37 0.6843847 -0.4781955  0.240057608 0.6953315 -0.5497993  0.27258080 0.7329225
#> 38 0.7445057 -0.4174942  0.075254931 0.7270802 -0.5099863  0.20775506 0.7033637
#> 39 0.7788422 -0.5099863 -0.023133445 0.7540568 -0.6452496  0.20182814 0.7957606
#> 40 0.7487194 -0.6284425  0.231117574 0.7669017 -0.7663123  0.25394672 0.7869693
#> 41 0.9245854 -1.0924038  0.680299776 0.9492859 -1.0804276  0.64700498 0.9136246
#> 42 0.9993115 -0.9365334  0.180465510 0.9990140 -1.0361829  0.25957349 0.9452113
#> 43 0.7321455 -0.8384155  0.304675592 0.7532961 -0.8534817  0.31475572 0.7884399
#> 44 1.0090822 -1.2125994  0.649540909 1.0143716 -1.1726473  0.60773348 0.9567649
#> 45 0.6835691 -1.0407491 -0.404087522 0.6491187 -1.0453363 -0.29305306 0.6299324
#> 46 0.7029636 -0.7945328  0.157256277 0.7282514 -0.8745755  0.23495881 0.7873371
#> 47 0.7483371 -0.5883480  0.151021411 0.7743812 -0.7838566  0.25957349 0.7740086
#> 48 0.6499629 -1.1069671 -0.117616837 0.6376512 -1.1292177 -0.04826039 0.6444626
#> 49 0.7069573 -0.5782103  0.340958097 0.7105379 -0.6498828  0.29026958 0.7278612
#> 50 0.7425844 -0.4028776  0.044038083 0.7395028 -0.5059567  0.22468254 0.7333108
#> 51 0.7795837 -0.5345106 -0.195087098 0.7758704 -0.6193926  0.02844930 0.7321455
#> 52 0.7364116 -0.7999140  0.367611220 0.7391169 -0.9181468  0.37207801 0.7563356
#> 53 0.7510104 -1.0226080  0.543664337 0.7570940 -1.1884371  0.52805749 0.7736358
#> 54 0.3552237 -0.9743530  0.110604709 0.3642480 -0.9786449  0.20380768 0.4044185
#> 55 0.7578519 -1.0047902  0.425763978 0.7627641 -1.0181237  0.42047016 0.7699002
#> 56 0.7972183 -1.1571030  0.647004982 0.7935700 -1.1726473  0.61212523 0.7961252
#> 57 0.6720794 -0.9851174 -0.314959742 0.6770197 -1.0025851 -0.18921049 0.6604562
#> 58 0.6949282 -0.9743530  0.008615253 0.6957347 -0.9872843  0.22597287 0.7093458
#> 59 0.7555766 -0.7663123  0.096027191 0.7601220 -0.9324182  0.26081960 0.7582306
#> 60 0.6478510 -1.2180490 -0.286573557 0.6444626 -1.1964265 -0.22804065 0.6448868
#> 61 0.7360245 -0.6284425  0.363124392 0.7239503 -0.7715433  0.36928860 0.7529155
#> 62 0.6807090 -0.5046171  0.173015425 0.7021630 -0.6299588  0.29689811 0.7429690
#> 63 0.6679438 -0.5912635  0.155874107 0.7045630 -0.7645746  0.30467559 0.7968541
#> 64 0.7728898 -0.8180633  0.420470160 0.8066421 -0.9060742  0.39629493 0.7968541
#> 65 0.8080840 -1.1118691  0.657110331 0.9002454 -1.0972347  0.62041679 0.8658226
#> 66 0.2044667 -1.0158891 -0.111273363 0.2253279 -1.0875961  0.09676114 0.3004952
#> 67 0.8277030 -0.9080762  0.444592986 0.8658226 -1.0136594  0.39738189 0.8177633
#> 68 0.9688002 -1.2262795  0.653332782 0.9229811 -1.1964265  0.63935839 0.9120026
#> 69 0.6921005 -1.1020891 -0.236195643 0.7081522 -1.1494205 -0.03394436 0.6550135
#> 70 0.6908862 -0.9894558  0.177085987 0.6929093 -1.0615591  0.36368635 0.7313679
#> 71 0.7517729 -0.7963233  0.177085987 0.7375719 -0.9365334  0.28663541 0.7635176
#> 72 0.6482737 -1.2800406 -0.203967275 0.6346565 -1.2429465 -0.04741258 0.6238875